add Movie detail and movie trailer. add pagination for movie list

+++ app/src/main/java/com/movies/model/common/movie/MovieModelCommon.kt
@@ -2,6 +2,8 @@ package com.movies.model.common.movie
 
 import com.movies.base.BaseCommonModel
 
-class MovieModelCommon(val id: Long,
-                       val name: String,
-                       val poster_path: String) : BaseCommonModel()
+class MovieModelCommon(
+    val id: Int,
+    val name: String,
+    val poster_path: String
+) : BaseCommonModel()
